feat: lift map function specific logic out of function closure resolve
