+++ lib/vrl/compiler/src/expression/function_closure.rs
@@ -1,6 +1,3 @@
-use diagnostic::{DiagnosticError, Label};
-use std::collections::btree_map::BTreeMap;
-
 use crate::expression::Block;
 use crate::parser::{Ident, Node};
 use crate::{Context, Expression, ExpressionError, Value};
@@ -25,7 +22,8 @@ impl FunctionClosure {
         &self,
         ctx: &mut Context,
         value: Value,
-        mut func: impl FnMut(&Context, Output, &mut Value) -> Result<(), ExpressionError>,
+        mut func: impl FnMut(&Context, Value, Value) -> Result<Value, ExpressionError>,
+        mut result: Value,
     ) -> Result<Value, ExpressionError> {
         match value {
             Value::Object(object) => {
@@ -42,33 +40,14 @@ impl FunctionClosure {
                     .clone()
                     .into_inner();
 
-                let mut result = Value::Object(BTreeMap::default());
                 for (key, value) in object.into_iter() {
                     let state = ctx.state_mut();
                     state.insert_variable(key_ident.clone(), key.into());
                     state.insert_variable(value_ident.clone(), value);
 
-                    let output = match self.block.resolve(ctx)? {
-                        Value::Array(mut array) => {
-                            let value = match array.pop() {
-                                Some(value) => Ok(value),
-                                None => Err(Error::ObjectArrayRequired.to_string()),
-                            }?;
+                    let output = self.block.resolve(ctx)?;
 
-                            let key = match array.pop() {
-                                Some(Value::Bytes(bytes)) => {
-                                    Ok(String::from_utf8_lossy(&bytes).into_owned())
-                                }
-                                None => Err(Error::ObjectArrayRequired.to_string()),
-                                _ => Err(Error::ObjectInvalidKey.to_string()),
-                            }?;
-
-                            Ok(Output::Object { key, value })
-                        }
-                        _ => Err(Error::ObjectNonArray.to_string()),
-                    }?;
-
-                    func(ctx, output, &mut result)?;
+                    result = func(ctx, output, result)?;
 
                     let state = ctx.state_mut();
                     state.remove_variable(&key_ident);
@@ -77,8 +56,6 @@ impl FunctionClosure {
                 Ok(result)
             }
             Value::Array(array) => {
-                let mut result = Value::Array(Vec::with_capacity(array.len()));
-
                 let index_ident = self
                     .variables
                     .get(0)
@@ -97,11 +74,9 @@ impl FunctionClosure {
                     state.insert_variable(index_ident.clone(), index.into());
                     state.insert_variable(value_ident.clone(), value);
 
-                    let output = Output::Array {
-                        element: self.block.resolve(ctx)?,
-                    };
+                    let output = self.block.resolve(ctx)?;
 
-                    func(ctx, output, &mut result)?;
+                    result = func(ctx, output, result)?;
 
                     let state = ctx.state_mut();
                     state.remove_variable(&index_ident);
@@ -149,25 +124,3 @@ pub enum Output {
 //         crate::TypeDef::new().null()
 //     }
 // }
-
-#[derive(thiserror::Error, Debug)]
-pub enum Error {
-    #[error("object iteration requires a two-element array return value")]
-    ObjectArrayRequired,
-
-    #[error("object iteration requires returning a key/value array return value")]
-    ObjectNonArray,
-
-    #[error("object iteration requires the first element to be a string type")]
-    ObjectInvalidKey,
-}
-
-impl DiagnosticError for Error {
-    fn code(&self) -> usize {
-        0
-    }
-
-    fn labels(&self) -> Vec<Label> {
-        vec![]
-    }
-}
