+++ lib/vrl/stdlib/Cargo.toml
@@ -43,6 +43,7 @@ roxmltree = { version = "0.14.1", optional = true }
 woothee = { version = "0.13.0", optional = true }
 uaparser = { version = "0.5.1", optional = true }
 utf8-width = { version = "0.1.5", optional = true }
+thiserror = "1"
 vector_common = { path = "../../vector-common", default-features = false, features = ["btreemap"] }
 
 [dev-dependencies]
